ferran melon

+++ Bicing/src/MainClasses/LocalSearchHeuristicFunction.java
@@ -14,7 +14,7 @@ import aima.search.framework.HeuristicFunction;
 public class LocalSearchHeuristicFunction implements HeuristicFunction {
 
     @Override
-    public double getHeuristicValue(Object o) {
+    public int getHeuristicValue(Object o) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
